Extract the analyser's hot paths into shadowable classes
MutatingScope carries code that is called millions of times per run: scope-table
merging and invalidation, conditional expression bookkeeping, expression keys,
and the yield scan. It is moved, unchanged, into plain classes that can be
reasoned about - and replaced - on their own:

- PHPStan\Analyser\ScopeOps
- PHPStan\Node\NodeScanner

The callers invoke them unconditionally; there is no branching and no behaviour
change.
