+++ phpstan-baseline.neon
@@ -69,7 +69,7 @@ parameters:
 		-
 			rawMessage: Casting to string something that's already string.
 			identifier: cast.useless
-			count: 3
+			count: 1
 			path: src/Analyser/MutatingScope.php
 
 		-
@@ -113,6 +113,12 @@ parameters:
 			identifier: new.internalClass
 			count: 1
 			path: src/Analyser/RuleErrorTransformer.php
+
+		-
+			rawMessage: Casting to string something that's already string.
+			identifier: cast.useless
+			count: 2
+			path: src/Analyser/ScopeOps.php
 
 		-
 			rawMessage: 'Doing instanceof PHPStan\Type\Constant\ConstantBooleanType is error-prone and deprecated. Use Type::isTrue() or Type::isFalse() instead.'
